backend/requirements.txt
fastapi>=0.115.0
sqlmodel>=0.0.20
asyncpg>=0.29.0
python-dotenv>=1.0.0
uvicorn>=0.29.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.25.0
PyJWT>=2.8.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0